LayoutTests/editing/caret/ios/emoji-expected.txt

PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS currentWidth is within 1 of 15
PASS successfullyParsed is true

TEST COMPLETE
